Coordinator/requirements.txt
Flask
mysql-connector-python==8.3.0